Moved IconView to icon/ component folder.
